+++ scripts/scripts.js
@@ -48,6 +48,24 @@ export function getLanguageCountryFromPath() {
   };
 }
 
+export function getOperatingSystem(userAgent) {
+  const systems = [
+    ['Windows NT 10.0', 'Windows 10'],
+    ['Windows NT 6.2', 'Windows 8'],
+    ['Windows NT 6.1', 'Windows 7'],
+    ['Windows NT 6.0', 'Windows Vista'],
+    ['Windows NT 5.1', 'Windows XP'],
+    ['Windows NT 5.0', 'Windows 2000'],
+    ['X11', 'X11'],
+    ['Mac', 'MacOS'],
+    ['Linux', 'Linux'],
+    ['Android', 'Android'],
+    ['like Mac', 'iOS'],
+  ];
+
+  return systems.find(([substr]) => userAgent.includes(substr))?.[1] || 'Unknown';
+}
+
 /**
  * Sets the page language.
  * @param {Object} param The language and country
